backend/requirements.txt
Flask
pymongo
flask-bcrypt
flask-cors
python-dotenv
